triage-pipeline/pyproject.toml
[project]
name = "ai-triage-pipeline"
version = "0.1.0"
description = "AI-powered intake & triage pipeline for a tutoring platform — Apex AI assessment"
authors = [{ name = "Roy Rizkallah" }]
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
    # --- Web / API ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "asgi-correlation-id>=4.3.4",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.12",

    # --- LangGraph / LangChain ---
    "langgraph>=0.2.50",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain>=0.3.7",
    "langchain-core>=0.3.18",

    # --- LLM gateway ---
    "litellm>=1.52.0",
    "instructor>=1.7.0",

    # --- Templating / prompts ---
    "Jinja2>=3.1.4",
    "PyYAML>=6.0.2",

    # --- Persistence ---
    "SQLAlchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg[binary,pool]>=3.2.3",
    "alembic>=1.14.0",

    # --- Queue ---
    "arq>=0.26.1",
    "redis>=5.2.0",

    # --- Observability ---
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-httpx>=0.49b0",
    "opentelemetry-instrumentation-asyncpg>=0.49b0",
    "opentelemetry-exporter-otlp>=1.28.0",

    # --- Utilities ---
    "httpx>=0.27.2",
    "tenacity>=9.0.0",
    "python-json-logger>=2.0.7",
    "rich>=13.9.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "respx>=0.21.1",
    "ruff>=0.7.4",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0.12",
    "syrupy>=4.7.2",
    "freezegun>=1.5.1",
    "ipython>=8.29.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N", "ASYNC", "S", "C4", "PIE", "RET", "SIM", "TCH"]
ignore = ["S101", "B008"]   # asserts ok in tests; FastAPI uses Depends() defaults

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "PLR2004"]
"alembic/**" = ["I", "E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]
exclude = ["alembic/versions/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "live: hits real LLM providers, requires API keys",
    "evals: gated eval suite (slower, blocks merge on regression)",
    "integration: requires postgres + redis",
]
filterwarnings = [
    "ignore::DeprecationWarning:pydantic.*",
]
